+++ src/Button/Button.test.js
@@ -8,7 +8,7 @@ describe('<Button />', () => {
             const element = mount(<Button data-sample='Sample' />);
 
             expect(
-                element.getDOMNode().attributes['data-sample'].value
+                element.find('.fd-button').getDOMNode().attributes['data-sample'].value
             ).toBe('Sample');
         });
     });
@@ -24,5 +24,28 @@ describe('<Button />', () => {
         mount(<Test />);
         expect(ref.current.tagName).toEqual('BUTTON');
         expect(ref.current.className).toEqual('fd-button');
+    });
+
+    describe('a11y', () => {
+        const element = mount(
+            <Button allowFocusOnDisable disabled
+                disabledMessage='To enable this button use the knob in the dev story'
+                enabledMessage='This button is now enabled'>Disabled Focusable</Button>
+        );
+        test('should have an aria-describedby attribute on the dialog', () => {
+            expect(
+                element.find('.fd-button').getDOMNode().attributes['aria-describedby'].value
+            ).toBeDefined();
+        });
+        test('should have an id in the instructions body', () => {
+            expect(
+                element.find('.fd-button__instructions').getDOMNode().attributes.id.value
+            ).toBeDefined();
+        });
+        test('should have an id on the instructions body that matches the button\'s aria-describedby', () => {
+            expect(
+                element.find('.fd-button').getDOMNode().attributes['aria-describedby'].value
+            ).toMatch(element.find('.fd-button__instructions').getDOMNode().attributes.id.value);
+        });
     });
 });